working on responsiveness of dashboard
