+++ src/App.css
@@ -62,3 +62,8 @@
 .sidebar.open ~ .css-hyum1k-MuiToolbar-root  {
   padding-left: 251px !important;
 }
+@media (max-width: 430px) {
+  .dashboard-wrapper {
+    padding-right: 1%;
+    padding-left: 1%;
+  }}
