+++ src/Components/Header/style.css
@@ -15,7 +15,7 @@
 .avatar {
   background-color: #fff;
   color: #1976d2;
-  margin-left: 20px;
+  margin-left: 70px;
   margin-right: 20px;
 }
 
@@ -23,12 +23,22 @@
   position: relative;
   border-radius: 4px;
   background-color: rgba(255, 255, 255, 0.15);
-  margin-left: 730px;
-  width: 100%;
-  max-width: 400px;
+  margin-left: 850px;
+  width: 80%;
+  max-width: 230px;
   display: flex;
 
   align-items: flex-end;
+}
+.searchIcon.mobile-only {
+  display: none;
+}
+/* .desktop-only {
+    display: flex;
+   
+} */
+.h1mobile-only{
+  display: none;
 }
 
 /* .searchIcon {
@@ -44,13 +54,18 @@
 }
 
 .inputInput {
-  padding: 8px 8px 8px 0;
-  padding-left: calc(1em + 32px);
-  transition: width 0.3s;
-  width: 100%;
+  padding: 8px 8px 8px 0 !important;
+  /* padding-left: calc(1em + 32px) !important; */
+  transition: width 0.3s !important;
+  width: 80%;
 }
 .css-hip9hq-MuiPaper-root-MuiAppBar-root{
   flex-direction: row !important;
+}
+.right-header{
+  display: flex;
+  flex-direction: row;
+  align-items: center;
 }
 
 /* @media (min-width: 600px) {
@@ -59,3 +74,55 @@
       padding-right: 24px;
   }
 } */
+@media (max-width: 430px) {
+  .css-hyum1k-MuiToolbar-root {
+      /* Adjustments for smaller screens */
+      padding-left: 0px !important;
+      padding-right: 0px !important;
+      min-height: 48px !important; /* or any other value that suits your design */
+      /* Add other adjustments here if needed */
+  }
+  .search {
+
+    margin-left: 0px;
+    margin-right: 0px;
+  }
+  .avatar {
+    /* background-color: #fff; */
+    /* color: #1976d2; */
+    margin-left: 0px;
+    margin-right: 0px;
+}
+.menuButton {
+  margin-right: 47% !important;
+  margin-left: 1% !important;
+}
+
+.searchIcon {
+  display: block;
+  padding: 0 16px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.search {
+  display: none;
+  
+}
+.searchIcon.mobile-only {
+  display: block;
+  padding-left: 70px;
+}
+
+.logoutIcon{
+  display: none !important;
+}
+/* Show search bar when the search icon is clicked */
+.searchOpen .search {
+  display: flex;
+}
+.h1mobile-only{
+  display: block;
+  
+}}
